resume-to-latex/pyproject.toml
[project]
name = "resume-to-latex"
version = "0.1.0"
description = "Convert resume from PDF or Word format to LaTeX"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "click>=8.1.7",
]

[project.scripts]
resume-to-latex = "resume_to_latex.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
